pyproject.toml
[project]
name = "simple-scheduler"
version = "0.1.0"
description = "Vista hackathon project"
requires-python = ">=3.12"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "httpx",
    "sqlalchemy",
    "claude-code-sdk",
    "django==5.2.4",
    "anthropic",
    "python-dotenv",
]
